+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 - Added a search field to the Collections panel
 - Added Drum Kit support to the standard N-SPC driver (Nintendo's SNES SDK driver).
 - Added Drum Kit support to the SuzukiSnes driver (Seiken Densetsu 3, Super Mario RPG, etc)
-- Added Drum Kit support to the KonamiSnes driver.
+- Added many improvements to the KonamiSnes driver.
 - Improved support for late-era Intelligent Systems SNES/SFC titles (Tetris Attack, Fire Emblem 4, Super Famicom Wars) 
 
 ### Changed